refactor: clean up code structure
