+++ lib/components/ChangelogList/MinimalChangelogList.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { useUpdateHiveContext } from '../ChangelogContext';
 import { ChangeTypeMap } from '../changelog.util.ts';
 import { ChangeType } from '../../changelog.types.ts';
-import SimpleList from './SimpleList.tsx';
+import SimpleList from './_internal/SimpleList.tsx';
 
 interface Props {
   changeTypeMapper?: Record<ChangeType, string>;
